addAsset now always returns a map containing directory information and files that were successfully uploaded and files that had issues uploading. Moved asset-related functions into new ProjectAssetAPIController. #2160
