+++ src/main/webapp/wise5/authoringTool/i18n/i18n_en.json
@@ -44,6 +44,7 @@
     "areYouSureYouWantToUploadThisLargeFileWhileUploadingMultipleFiles": "One of the files you are trying to upload is larger than 500 KB. We recommend using smaller files so they load faster for students. Are you sure you want to upload the large file?",
     "areYouSureYouWantToUploadTheseLargeFilesWhileUploadingMultipleFiles": "{{fileCount}} of the files you are trying to upload are larger than 500 KB. We recommend using smaller files so they load faster for students. Are you sure you want to upload these large files?",
     "assetUploadSuccessful": "Successfully uploaded: {{assetFilenames}}",
+    "assetUploadError": "Error uploading the following files: ",
     "assetUsageMessage": "You are using {{usage}} out of {{max}} ({{percentage}}%)",
     "authoringTool": "Authoring Tool",
     "authoringToolMenu": "Authoring Tool Menu",
